Added a row to allow users to insert new cities, occupations, relationships, shortened some code, now working on changing the cursor when hovered over an editable textbox.

+++ rmanager/editForm.cs
@@ -35,6 +35,8 @@ namespace rmanager
                     break;
             }
             setDataGridValues(name);
+            editDataGridView.Height = 179;
+            this.Height = 218;
         }
         
         private void setDataGridValues(string name)
@@ -79,10 +81,15 @@ namespace rmanager
                 editFormDataGridRecordBindingSource.Add(new editFormDataGridRecord()
                 {
                     Id = (int)dt.Rows[i]["id"],
-                    Value = (string)dt.Rows[i][column_name]
+                    Value = Utilities.CapitalizeFirstLetters((string)dt.Rows[i][column_name])
                 });
             }
-           
+
+            editFormDataGridRecordBindingSource.Add(new editFormDataGridRecord()
+            {
+                Id = dt.Rows.Count,
+                Value = ""
+            });  
         }
         //editDataGridView.Rows[e.RowIndex].Cells[1].Value.ToString()
         private void editDataGridView_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -104,15 +111,42 @@ namespace rmanager
             int city_id = (int) editDataGridView.Rows[e.RowIndex].Cells[1].Value;
             if (e.ColumnIndex == 3)
             {
-                if (MessageBox.Show(message,
-                                   "WARNING!",
-                                   MessageBoxButtons.YesNo,
-                                   MessageBoxIcon.Warning) == DialogResult.Yes)
+                if(editDataGridView.Rows[e.RowIndex].Cells[e.ColumnIndex].Value.ToString() == "Delete")
+                { 
+                    if (MessageBox.Show(message,
+                                       "WARNING!",
+                                       MessageBoxButtons.YesNo,
+                                       MessageBoxIcon.Warning) == DialogResult.Yes)
+                    {
+                        Utilities.MySqlCommandImproved($"DELETE FROM user_{name} WHERE user_id = {user_id} AND {column_name}_id = {city_id}");
+                        editDataGridView.Rows.RemoveAt(e.RowIndex);
+                    }
+                }else
                 {
-                    Utilities.MySqlCommandImproved($"DELETE FROM user_{name} WHERE user_id = {user_id} AND {column_name}_id = {city_id}");
-                    editDataGridView.Rows.RemoveAt(e.RowIndex);
+                    this.Close();
                 }
+                
             }
+            Point pt = Cursor.Position;
+            MessageBox.Show(pt.X + " " + pt.Y);
+        }
+
+        private void editDataGridView_DataBindingComplete(object sender, DataGridViewBindingCompleteEventArgs e)
+        {
+            for (int i = 0; i < editDataGridView.Rows.Count - 1; i++)
+            {
+                editDataGridView.Rows[i].Cells[2].Value = "Edit";
+                editDataGridView.Rows[i].Cells[3].Value = "Delete";
+            }
+            editDataGridView.Rows[editDataGridView.Rows.Count - 1].Cells[0].ReadOnly = false;
+            editDataGridView.Rows[editDataGridView.Rows.Count - 1].Cells[2].Value = "Add New";
+            editDataGridView.Rows[editDataGridView.Rows.Count - 1].Cells[3].Value = "Exit";
+        }
+
+        private void editDataGridView_MouseHover(object sender, EventArgs e)
+        {
+            
+            //MessageBox.Show(GetChildAtPoint(Cursor.Position).ToString());
         }
     }
 }
